moving Tree Diagram to its own folder with its style sheet
